+++ tests/mocks/usePlayerStore.ts
@@ -1,6 +1,0 @@
-import { vi } from "vitest";
-
-vi.mock('@/stores/PlayerStore', () => ({
-    usePlayerStore: () => ({
-    })
-}));